add auth endpoint
